Move invoke class to criterion model and subclass domain classes for each test.
